Move Linux-specific `URL` extensions into `PlatformSupportLinux`
