+++ Sources/InContextCore/Extensions/URL.swift
@@ -73,46 +73,3 @@ extension URL {
     }
 
 }
-
-#if os(Linux)
-
-enum DirectoryHint {
-
-    case isDirectory
-    case notDirectory
-
-}
-
-extension URL {
-
-    init(filePath: String, directoryHint: DirectoryHint = .notDirectory) {
-        self.init(fileURLWithPath: filePath, isDirectory: directoryHint == .isDirectory ? true : false)
-    }
-
-    init(filePath: String, relativeTo url: URL?) {
-        self.init(fileURLWithPath: filePath, relativeTo: url)
-    }
-
-    var pathIncludingTrailingDirectorySeparator: String {
-	if hasDirectoryPath {
-            return path + "/"
-        }
-        return path
-    }
-
-    func relative(to url: URL) -> URL {
-        precondition(isFileURL)
-        precondition(url.isFileURL)
-	precondition(url.hasDirectoryPath)
-        let directoryPath = url.pathIncludingTrailingDirectorySeparator
-        let path = pathIncludingTrailingDirectorySeparator
-        precondition(path.starts(with: directoryPath))
-	let relativePath = String(path.dropFirst(directoryPath.count))
-	print(relativePath)
-
-        return URL(filePath: relativePath, relativeTo: url)
-    }
-
-}
-
-#endif
